fix(profiling): fix potential crash in frame converter (DataDog/dd-trace-py#4119)
On Python 3.9 and later, PyFrame_GetCode returns a new reference that should be
decremented after the function finishes.

The current code calls Py_DECREF without checking that code is not NULL, while
there is a check doing that for safety. The DECREF is also done a little too
soon (in theory, in practice it does not have a real impact).

This switches to XDECREF to make sure that `code` is not NULL and does this at
the end of the function to be sure we have inc-referenced code items that we
cared about.

Vendor-Component: memalloc
Upstream-Commit: 7f0943a0ff2d5f93a4ca4125632729319887fb14