feat(profiling): extract class name from stack traces (DataDog/dd-trace-py#4012)
This adds an heuristic for extracting class names from stack traces. This is
far from being perfect, as it relies on the convention of using `self` and
`cls` as name for the local variable in a function call to designate instances
or classes.

However, that ought to be enough for the vast majority of the Python program
seen in the wild.

Vendor-Component: memalloc
Upstream-Commit: e8681f7fdf28849885686444eee03e9537160482