refactor(profiling): add more GIL assertions to memalloc (DataDog/dd-trace-py#12000)
The first GIL assertion I added, to memalloc_add_event, has not tripped
yet on a test application. One the one hand, it's reassuring that we
always see the GIL in that part of the code. On the other hand, there
are other parts of the memory profiler that could in theory be called
concurrently where I didn't add the assertion: when stopping the
profiler, and when iterating over the events to aggregate them. Add GIL
assertions to those points.  The goal is ultimately to understand why we
needed to add locks to the profiler to prevent it from crashing, given
that the GIL exists.

Vendor-Component: memalloc
Upstream-Commit: ef4c99759f8c1d87642718db2292ae98e073b3c6