refactor(profiling): add accuracy tests for heap profiling (DataDog/dd-trace-py#12394)
There aren't any tests of the statistical accuracy of the heap profiler.
This PR adds two tests.

One test allocates known amounts of memory in different functions, keeps
those allocations alive so they'll be captured in a heap profile, and
compares the values reported in the heap profile to the actual amount of
memory allocated according to tracemalloc. This is repeated for several
different sampling rates. The test makes loose assertions about the
amonut of live memory measured in the profile and the relative
proportion of memory used by specific functions. The assertions could be
tighter but they'd likely make the test flaky.

The second test is a simplified version of the first, intended to be run
with high sampling, that will be really really slow with our current
implementation. The test program used to crash. It doesn't now, but it's
too slow to be worth running in CI. But one goal of my work is to make
sure the overhead of the profiler is reasonable, especially the scaling
behavior for tracking lots of allocations. If the performance improves,
we can run it as a regression test.

PROF-11313

Vendor-Component: memalloc
Upstream-Commit: 72b140d835d2ea4cedad07f4e326a5790af90bae