refactor(profiling): pool memalloc traceback scratch buffers (DataDog/dd-trace-py#13185)
Right now there is a global scratch buffer used when collecting
tracebacks for memalloc. This is a resource usage optimization. We have
enough space set aside to record a maximum depth traceback. We collect
the traceback and put it in the scratch space. Then, once we know how
many frames there really are, we can allocate a new traceback that's
just the right size and store that for later reporting. This saves
wasted memory and extra work dynamically growing the traceback to the
right size.

However, there is only one scratch buffer. There is a hypothetical
logical race accessing the buffer. Before 2.18, when we added locks, it
was a _real_ logical  race. Essentially what happens is:

- Thread A is going to collect a traceback, and starts using the buffer
- Thread A sets the frame count to 0 and starts writing frames to the
  buffer
- Thread A, partway through, triggers a memory allocation when getting
  the next frame that just so happens to trigger garbage collection,
  which in turn just so happens to run interpreter code which releases
  the GIL
- Thread B can start running now that the GIL was released
- Thread B happens to also want to collect a traceback. It can do so and
  starts using the same buffer
- Thread B resets the frame count to 0
- Thread B writes all its frames, leaving the frame count at N, and just
  so happens _not_ to relase the GIL
- Thread B finishes sampling and eventually releases the GIL
- Thread A picks up where it left off, continuing to write frames. Only
  now the frame counter is N, so it is writing at the Nth offset in the
  buffer.
- Thread A runs out of Python frames, and returns a new traceback, which
  will have the N frames that thread B wrote plus the remaining frames
  that A collected.

This is a logical race that leads to double frees. Now we have two
tracebacks holding some of the exact same frame objects, and they will
eventually both be freed. This corrupts internal allocator state and
leads to crashes.

I say this is hypothetical because now, even though we have locks in
this code, there are actually two separate locks which can be acquired
independently. Both the allocation profiler and heap profiler could
simultaneously collect tracebacks. I haven't been able to actually
trigger this, even with my recently-added reproducer for the race that
ddtrace 2.18 addressed, but I believe the problem is there.

Even without proving this is a problem in practice, it is also an
impedment to performance optimizations. The synchronization we added in
ddtrace 2.18 has introduced a lof of overhead. We can, in theory, use
the GIL to protect most of the memory profiler data structures. But we
need to guard against logical races like this.

Keeping one buffer and locking it for the duration of traceback
collection is not a solution. If the GIL could possibly be released,
then this creates a lock order inversion bug. Another thread grabs the
GIL and tries to get a traceback, but can't since the lock is held. And
so the thread can't release the GIL. So the program deadlocks. For the
same reason, we can't just give the allocation and heap profilers their
own scratch buffers.

NB: We are actually only saved from this problem today because we use
try locks. And I'm not confident that we are actually correctly handling
the cases where fail to acquire the existing locks. I would rather have
a solution where we don't possibly introduce lock ordering bugs.

Instead, we can use a pool of buffers. Basically, when a thread wants a
traceback, it grabs a buffer from the pool or creates a new one when the
pool is empty. Then the thread has exclusive access to that buffer. When
the thread is done with the buffer, it places it back in the pool. We
can cap the size of the pool to prevent having the pool stuck at the
high water-mark size if we typically only need one or two buffers at a
time.

This is safe if we hold the GIL on entry to the buffer management
functions, as the functions themselves don't call back into Python and
risk releasing the GIL. If we ever wanted to support no-GIL mode or
subinterpreters, wit would be trivial to add synchronization to the
pool.

I suspect that in the stable state we will not need to allocate new
buffers often and the overhead will be comparable to what it is today.
But it'll be much safer and more suitable for further optimizations to
the memory profiler.

PROF-11531

Vendor-Component: memalloc
Upstream-Commit: 8c189a16fd8ebaaf51b83531a06d728dce9dc43b