fix(profiling): actually disable GC during memalloc traceback collection (DataDog/dd-trace-py#15388)
This corrects a previous, invalid fix for a use-after-free crash in the
memory profiler (DataDog/dd-trace-py#14550). The previous fix used preprocessor macros to
conditionally disable/reenable GC for Python 3.10 and 3.11. However, the
previous fix failed to include `_pymacro.h`, which defines the macros.
So, the fix didn't actually apply.

Vendor-Component: memalloc
Upstream-Commit: 641355b41e7f3ce5002e686e5d986136b8bbd5d9